requirements.txt
# Core ML/DL Dependencies - CUDA Enabled
torch>=2.0.0+cu121
torchvision+cu121
torchaudio+cu121

transformers
diffusers>=0.30.0
accelerate
timm

# Wavelet Transform
pytorch-wavelets
PyWavelets

# Data & Visualization
numpy
matplotlib
seaborn
pillow
datasets

# Evaluation Metrics
pytorch-fid
torch-fidelity
torcheval
torchmetrics
clean-fid

# Utility
tqdm
einops
pyyaml

# Jupyter/Development (optional)
ipython
ipykernel
jupyter

# Tensorboard for logging
tensorboard
